pyproject.toml
[tool.poetry]
name = "pythonprojectdjango"
version = "0.1.0"
description = ""
authors = ["Максим Ландау <12landay@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
django-rest-framework = "^0.1.0"
django-filter = "^23.5"
django-debug-toolbar = "^4.3.0"
pillow = "^10.2.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
